[Shader] Add #include to shader

+++ engine/compute_shader.cc
@@ -77,7 +77,7 @@ void ComputeShader::SetScreenSize(const glm::vec2& screen_size) {
 }
 
 void ComputeShader::SetTimeSeed() {
-  shader_.SetFloat("time_for_seed", util::AsFloat(util::Now()));
+  shader_.SetInt("time_for_seed", util::AsInt(util::Now()));
 }
 
 };
